Split DefaultCloudService into DefaultTemplateCloudService
